--- projects/angular-split/src/lib/interface/IArea.ts
@@ -1,9 +1,0 @@
-import { SplitAreaDirective } from "../directive/splitArea.directive";
-
-export interface IArea {
-    component: SplitAreaDirective;
-    order: number;
-    size: number;
-    minSize?: number;
-    maxSize?: number;
-}
