--- projects/angular-split/src/lib/interface/IAreaSnapshot.ts
@@ -1,7 +1,0 @@
-import { IArea } from "./IArea";
-
-export interface IAreaSnapshot {
-    area: IArea
-	sizePixelAtStart: number
-	sizePercentAtStart: number
-}
