--- projects/angular-split/src/lib/interface/IPoint.ts
@@ -1,5 +1,0 @@
-
-export interface IPoint {
-    x: number;
-    y: number;
-}
