--- projects/angular-split/src/lib/interface/ISplitSnapshot.ts
@@ -1,9 +1,0 @@
-import { IAreaSnapshot } from './IAreaSnapshot'
-
-export interface ISplitSnapshot {
-	gutterNum: number
-	containerSizePixel: number
-    lastSteppedOffset: number
-    areasBeforeGutter: Array<IAreaSnapshot>
-    areasAfterGutter: Array<IAreaSnapshot>
-}
